+++ src/main/java/gtPlusPlus/core/item/base/BaseItemComponent.java
@@ -303,8 +303,8 @@ public class BaseItemComponent extends Item {
                 ((Map<Integer, Short[]>) extraData).computeIfAbsent(9999, k -> new Short[] {0});
                 Short aCurrentFrame = ((Map<Integer, Short[]>) extraData).get(9999)[0];
                 short fC = (short) (aCurrentFrame >= Short.MAX_VALUE ? 0 : aCurrentFrame + 1);
-                ((Map<Integer, Short[]>) extraData).put((int) 9999, new Short[] {(short) (fC), 0});
-                ((Map<Integer, Short[]>) extraData).put((int) 9998, new Short[] {aCurrentFrame, 0});
+                ((Map<Integer, Short[]>) extraData).put(9999, new Short[] {fC, 0});
+                ((Map<Integer, Short[]>) extraData).put(9998, new Short[] {aCurrentFrame, 0});
             }
         }
     }
